Fixed compile options in cmake
Fixed a lot of conversion warnings caused by robust compiler options

+++ src/graphics/TouchParser.cpp
@@ -120,7 +120,7 @@ void TouchParser::readBody(TouchEvent &arInput)
         readRawTouchEvent();
     }
 
-    std::uint64_t m = (mRawTouchEvent.stime * 1000) + (mRawTouchEvent.mtime / 1000);
+    std::uint64_t m = (mRawTouchEvent.stime * 1000) + static_cast<uint32_t>(mRawTouchEvent.mtime / 1000);
     arInput.mTime = std::chrono::steady_clock::time_point(std::chrono::milliseconds{m});
 }
